Move all built-ins to a dedicated file
