+++ src/builtin_bindings.h
@@ -1,0 +1,8 @@
+#ifndef __BUILTIN_BINDINGS_H__
+#define __BUILTIN_BINDINGS_H__
+
+#include "duktape/duktape.h"
+
+void init_builtins(duk_context *ctx);
+
+#endif
